% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_model.R
\name{agg_curve}
\alias{agg_curve}
\title{Aggregate Learning Curve}
\usage{
agg_curve(t, r, n, na.rm = FALSE)
}
\arguments{
\item{t}{vector of hours (or costs) for the first unit from departments 1 through m}

\item{r}{vector of historical learning rates for departments 1 through m}

\item{n}{total units to be produced across all departments}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Computes the approximate aggregate cumulative learning curve formula by calculating the sum of all contributing hours from all departments for all production units 1 through n.
}

