% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_avg_models.R
\name{ca_block}
\alias{ca_block}
\title{Wright's Cumulative Average Learning Curve Function}
\usage{
ca_block(t, n, r, m = 1, na.rm = FALSE)
}
\arguments{
\item{t}{time (or cost) required for the mth unit of production}

\item{n}{last unit of the production block of concern}

\item{r}{learning curve rate}

\item{m}{first unit of the production block of concern (default: m = 1)}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Computes cumulative time or cost for units m through n in a production block using Wright's cumulative average model. Assumes the block begins at unit m and ends at unit n.
}

