% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_and_error_models.R
\name{delta}
\alias{delta}
\title{Crawford vs. Wright Unit Difference}
\usage{
delta(t, m, n, r, level = "u")
}
\arguments{
\item{t}{time (or cost) required to produce the first unit}

\item{m}{mth unit for which you have time (or cost) information (default is m = 1)}

\item{n}{the nth unit you wish to predict the time (or cost) for when comparing unit predictions or the last unit in the block when comparing cumulative time (or costs)}

\item{r}{learning curve rate}

\item{level}{calculate unit ("u") versus cumulative ("c") differences (default = "u")}
}
\description{
Computes the difference between the unit or cumulative prediction estimates provided by the Crawford and Wright models.
}

