% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLearningCurve.R
\name{generateLearningCurve}
\alias{generateLearningCurve}
\title{generateLearningCurve}
\usage{
generateLearningCurve(workFolder)
}
\arguments{
\item{workFolder}{the folder containing the data and populations}
}
\description{
generateLearningCurve generates learning curves for all target data in the
work folder.
}
\examples{
# register a parallel backend
PatientLevelPrediction::registerParallelBackend()

# generate the learning curves
generateLearningCurve(workFolder)

# de-register parallel backend by registering a sequential backend
PatientLevelPrediction::registerSequentialBackend()
}
