% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_rate_models.R
\name{lc_rate}
\alias{lc_rate}
\title{Learning Rate Converter}
\usage{
lc_rate(b, na.rm = FALSE)
}
\arguments{
\item{b}{natural slope}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Computes the learning rate for given natural slopes
}

