% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_rate_models.R
\name{natural_slope}
\alias{natural_slope}
\title{Natural Slope Rate Converter}
\usage{
natural_slope(r, na.rm = FALSE)
}
\arguments{
\item{r}{learning curve rate}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Computes the natural slope rate for given learning rates
}
\examples{
# Calculate the natural slope for learning rates of 80\%, 85\%, 90\%

natural_slope(r = c(.80, .85, .90))
## [1] -0.3219281 -0.2344653 -0.1520031

}
