% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitResults.R
\name{removeSensitive}
\alias{removeSensitive}
\title{Remove any sensitive details before submitting results}
\usage{
removeSensitive(workFolder, N = NULL, localFolder = NULL)
}
\arguments{
\item{workFolder}{Name of results folder}

\item{N}{The minimum count size for summary data to be exported}

\item{localFolder}{Name of the folder where senitive data gets moved}
}
\description{
Remove any sensitive details before submitting results
}
\details{
This function removes any passwords/usernames in the connectionDetails
}
