% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_models.R
\name{unit_cum_appx}
\alias{unit_cum_appx}
\title{Approximate Cumulative Unit Learning Curve Function}
\usage{
unit_cum_appx(t, n, r, m = 1, na.rm = FALSE)
}
\arguments{
\item{t}{time (or cost) required for the mth unit of production}

\item{n}{The unit you wish to predict the cumulative time (or cost) to}

\item{r}{learning curve rate}

\item{m}{mth unit of production (default set to 1st production unit)}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Provides the approximate cumulative time or cost required for units m through n (inclusive) using the Crawford unit model. Provides nearly the exact output as unit_cum_exact(), usually only off by 1-2 units but reduces computational time drastically if trying to calculate cumulative hours (costs) for over a million units.
}

