% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_models.R
\name{unit_curve}
\alias{unit_curve}
\title{Crawford's Unit Learning Curve Function}
\usage{
unit_curve(t, n, r, m = 1, na.rm = FALSE)
}
\arguments{
\item{t}{time (or cost) required for the mth unit of production}

\item{n}{nth unit you wish to predict the time (or cost) for}

\item{r}{learning curve rate}

\item{m}{mth unit of production (default set to 1st production unit)}

\item{na.rm}{Should \code{NA} values be removed?}
}
\description{
Predicts the time or cost of the nth unit given the time of the mth unit and the learning rate
}

