% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wright.R
\name{wright}
\alias{wright}
\title{Cumulative Average Learning Curve Production Function}
\usage{
wright(a = NULL, ...)
}
\arguments{
\item{a}{a number}

\item{...}{extra stuff}
}
\description{
does really cool stuff
}

