% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_rate_models.R
\name{natural_slope_est}
\alias{natural_slope_est}
\title{Natural Slope Estimate}
\usage{
natural_slope_est(T, t, n)
}
\arguments{
\item{T}{total time (or cost) required to produce the first n units}

\item{t}{time (or cost) required to produce the first unit}

\item{n}{total n units produced}
}
\description{
Computes the natural slope based on total time (cost) to produce the first n units, time (cost) required for the first unit and total units produced.
}

