% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_delta}
\alias{plot_delta}
\title{Crawford vs. Wright Delta Plot}
\usage{
plot_delta(t, m, n, r, level = "u")
}
\arguments{
\item{t}{time (or cost) required to produce the mth unit}

\item{m}{mth unit for which you have time (or cost) information (default is m = 1)}

\item{n}{the nth unit you wish to predict the time (or cost) for when comparing unit predictions or the last unit in the block when comparing cumulative time (or costs)}

\item{r}{learning curve rate}

\item{level}{plot the delta between the Crawford and Wright models at the unit ("u") or cumulative ("c") level}
}
\description{
Plots the delta of hours (or cost) per unit between Crawford's unit model and Wright's cumulative average model.
}

